pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "waluigi"
version = "0.1.0"
dependencies = [
    "requests",
    "tabulate",
    "ConfigArgParse",
    "pyyaml",
    "fastapi",
    "uvicorn",
    "httpx"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["waluigi*"]
namespaces = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[project.scripts]
wlctl = "waluigi.cli:main"
wlboss = "waluigi.boss.__main__:main"
wlworker = "waluigi.worker:main"
wlcatalog = "waluigi.catalog.__main__:main"

